intro modal to warn about zkey size if not connected to wifi
