+++ app/package.json
@@ -15,6 +15,7 @@
     "@ethersproject/shims": "^5.7.0",
     "@react-native-async-storage/async-storage": "^1.23.1",
     "@react-native-community/clipboard": "^1.5.1",
+    "@react-native-community/netinfo": "^11.3.1",
     "@tamagui/colors": "^1.94.3",
     "@tamagui/config": "^1.94.3",
     "@tamagui/core": "^1.94.3",
