+++ app/src/screens/MainScreen.tsx
@@ -10,13 +10,14 @@ import ProveScreen from './ProveScreen';
 import { Steps } from '../utils/utils';
 import AppScreen from './AppScreen';
 import { App } from '../utils/AppClass';
-import { Linking, Platform, Pressable } from 'react-native';
+import { Linking, Modal, Platform, Pressable } from 'react-native';
 import { Keyboard } from 'react-native';
 import NFC_IMAGE from '../images/nfc.png'
 import { bgColor, blueColorLight, borderColor, componentBgColor, textColor1, textColor2 } from '../utils/colors';
 import MintScreen from './MintScreen';
-import { ToastViewport } from '@tamagui/toast';
+import { ToastViewport, useToastController } from '@tamagui/toast';
 import { ToastMessage } from '../components/ToastMessage';
+import { downloadZkey } from '../utils/zkeyDownload';
 
 interface MainScreenProps {
   onStartCameraScan: () => void;
@@ -43,6 +44,9 @@ interface MainScreenProps {
   majority: number;
   setMajority: (age: number) => void;
   zkeydownloadStatus: string;
+  showWarning: boolean;
+  setShowWarning: (value: boolean) => void;
+  setDownloadStatus: (value: "not_started" | "downloading" | "completed" | "error") => void;
 }
 
 const MainScreen: React.FC<MainScreenProps> = ({
@@ -69,7 +73,10 @@ const MainScreen: React.FC<MainScreenProps> = ({
   setDateOfExpiry,
   majority,
   setMajority,
-  zkeydownloadStatus
+  zkeydownloadStatus,
+  showWarning,
+  setShowWarning,
+  setDownloadStatus
 }) => {
   const [NFCScanIsOpen, setNFCScanIsOpen] = useState(false);
   const [SettingsIsOpen, setSettingsIsOpen] = useState(false);
@@ -79,6 +86,8 @@ const MainScreen: React.FC<MainScreenProps> = ({
   const [selectedApp, setSelectedApp] = useState<App | null>(null);
   const [brokenCamera, setBrokenCamera] = useState(false);
   const [hideData, setHideData] = useState(false);
+
+  const toast = useToastController();
 
   const handleRestart = () => {
     setStep(Steps.MRZ_SCAN);
@@ -162,214 +171,142 @@ const MainScreen: React.FC<MainScreenProps> = ({
   const { height, width } = useWindowDimensions();
 
   return (
-    <YStack f={1} bc="#161616" mt={Platform.OS === 'ios' ? "$8" : "$0"} >
-      <YStack >
-        <XStack jc="space-between" ai="center" px="$3">
+    <>
+      <YStack f={1} bc="#161616" mt={Platform.OS === 'ios' ? "$8" : "$0"} >
+        <YStack >
+          <XStack jc="space-between" ai="center" px="$3">
 
-          <Button p="$2" py="$3" pr="$7" unstyled onPress={() => setSettingsIsOpen(true)}><Cog color="#a0a0a0" /></Button>
-
-
-
-          <Text fontSize="$6" color="#a0a0a0">
-            {selectedTab === "scan" ? "Scan" : (selectedTab === "app" ? "Apps" : "Prove")}
-          </Text>
-
-          <Button p="$2" py="$3" pl="$7" unstyled onPress={() => setHelpIsOpen(true)}><HelpCircle color="#a0a0a0" /></Button>
-
-
-        </XStack>
-        <Sheet open={NFCScanIsOpen} onOpenChange={setNFCScanIsOpen} modal dismissOnOverlayPress={false} disableDrag animation="medium" snapPoints={[35]}>
-          <Sheet.Overlay />
-          <Sheet.Frame>
-            <YStack gap="$5" f={1} pt="$3">
-              <H2 textAlign='center'>Ready to scan</H2>
-              {step >= Steps.NFC_SCAN_COMPLETED ?
-                <CheckCircle2
-                  size="$8"
-                  alignSelf='center'
-                  color="#3185FC"
-                  animation="quick"
-                /> :
-                <Image
-                  h="$8"
-                  w="$8"
-                  alignSelf='center'
-                  borderRadius={1000}
-                  source={{
-                    uri: NFC_IMAGE
-                  }}
-                />
-              }
-              <Text textAlign='center'>Hold your device near the NFC tag.</Text>
-            </YStack>
-          </Sheet.Frame>
-        </Sheet>
-
-        <Sheet open={SettingsIsOpen} onOpenChange={setSettingsIsOpen} modal animation="medium" snapPoints={[88]}>
-          <Sheet.Overlay />
-          <Sheet.Frame bg={bgColor} borderRadius="$9">
-            <YStack p="$3" pb="$5" f={1} gap={height > 750 ? "$3" : "$1"} mb="$1.5">
-              <XStack gap="$2" ml="$2" >
-                <H2 color={textColor1}>Settings</H2>
-                <Cog color={textColor1} mt="$1" alignSelf='center' size="$2" />
-              </XStack>
-              <Fieldset horizontal>
-                <Label color={textColor1} width={225} justifyContent="flex-end" htmlFor="name" >
-                  Broken camera
-                </Label>
-                <Switch size="$3.5" checked={brokenCamera} onCheckedChange={setBrokenCamera}>
-                  <Switch.Thumb animation="bouncy" bc={bgColor} />
-                </Switch>
-              </Fieldset>
-              {
-                brokenCamera &&
-                <YStack pl="$3" gap="$1">
-                  <Fieldset gap="$4" horizontal>
-                    <Label color={textColor1} width={160} justifyContent="flex-end" fontSize={13}>
-                      Passport Number
-                    </Label>
-                    <Input bg={componentBgColor} color={textColor1} h="$3.5" borderColor={passportNumber?.length === 9 ? "green" : "unset"} flex={1} id="passport_number" onChangeText={(text) => setPassportNumber(text.toUpperCase())} value={passportNumber} keyboardType="default" />
-                  </Fieldset>
-                  <Fieldset gap="$4" horizontal>
-                    <Label color={textColor1} width={160} justifyContent="flex-end" fontSize={13}>
-                      Date of birth (yymmdd)
-                    </Label>
-                    <Input bg={componentBgColor} color={textColor1} h="$3.5" borderColor={dateOfBirth?.length === 6 ? "green" : "unset"} flex={1} id="date_of_birth" onChangeText={setDateOfBirth} value={dateOfBirth} keyboardType={Platform.OS == "ios" ? "default" : "number-pad"} />
-                  </Fieldset>
-                  <Fieldset gap="$4" horizontal>
-                    <Label color={textColor1} width={160} justifyContent="flex-end" fontSize={13}>
-                      Date of expiry (yymmdd)
-                    </Label>
-                    <Input bg={componentBgColor} color={textColor1} h="$3.5" borderColor={dateOfExpiry?.length === 6 ? "green" : "unset"} flex={1} id="date_of_expiry" onChangeText={setDateOfExpiry} value={dateOfExpiry} keyboardType={Platform.OS == "ios" ? "default" : "number-pad"} />
-                  </Fieldset>
-                </YStack>
-              }
-
-              <Fieldset horizontal>
-                <Label color={textColor1} width={225} justifyContent="flex-end" htmlFor="restart" >
-                  Private mode
-                </Label>
-                <Switch size="$3.5" checked={hideData} onCheckedChange={handleHideData}>
-                  <Switch.Thumb animation="bouncy" bc={bgColor} />
-                </Switch>
-              </Fieldset>
+            <Button p="$2" py="$3" pr="$7" unstyled onPress={() => setSettingsIsOpen(true)}><Cog color="#a0a0a0" /></Button>
 
 
 
-              <Fieldset gap="$4" mt="$1" horizontal>
-                <Label color={textColor1} width={200} justifyContent="flex-end" htmlFor="restart">
-                  Restart to step 1
-                </Label>
-                <Button bg={componentBgColor} jc="center" borderColor={borderColor} borderWidth={1.2} size="$3.5" ml="$2" onPress={handleRestart}>
-                  <IterationCw color={textColor1} />
-                </Button>
-              </Fieldset>
+            <Text fontSize="$6" color="#a0a0a0">
+              {selectedTab === "scan" ? "Scan" : (selectedTab === "app" ? "Apps" : "Prove")}
+            </Text>
 
-              <Fieldset gap="$4" mt="$1" horizontal>
-                <Label color={textColor1} width={200} justifyContent="flex-end" htmlFor="skip" >
-                  Use mock passport data
-                </Label>
-                <Button bg={componentBgColor} jc="center" borderColor={borderColor} borderWidth={1.2} size="$3.5" ml="$2" onPress={handleSkip}>
-                  <VenetianMask color={textColor1} />
-                </Button>
-              </Fieldset>
-              <YStack flex={1} />
+            <Button p="$2" py="$3" pl="$7" unstyled onPress={() => setHelpIsOpen(true)}><HelpCircle color="#a0a0a0" /></Button>
 
-              <YStack mb="$0">
-                <Button p="$2.5" borderRadius="$3" bg={componentBgColor} jc="center" borderColor={borderColor} borderWidth={1.2} onPress={() => setSettingsIsOpen(false)} w="80%" alignSelf='center'>
-                  <Text color={textColor1} textAlign='center' fow="bold">Close</Text>
-                </Button>
+
+          </XStack>
+          <Sheet open={NFCScanIsOpen} onOpenChange={setNFCScanIsOpen} modal dismissOnOverlayPress={false} disableDrag animation="medium" snapPoints={[35]}>
+            <Sheet.Overlay />
+            <Sheet.Frame>
+              <YStack gap="$5" f={1} pt="$3">
+                <H2 textAlign='center'>Ready to scan</H2>
+                {step >= Steps.NFC_SCAN_COMPLETED ?
+                  <CheckCircle2
+                    size="$8"
+                    alignSelf='center'
+                    color="#3185FC"
+                    animation="quick"
+                  /> :
+                  <Image
+                    h="$8"
+                    w="$8"
+                    alignSelf='center'
+                    borderRadius={1000}
+                    source={{
+                      uri: NFC_IMAGE
+                    }}
+                  />
+                }
+                <Text textAlign='center'>Hold your device near the NFC tag.</Text>
               </YStack>
-            </YStack>
-          </Sheet.Frame>
-        </Sheet>
+            </Sheet.Frame>
+          </Sheet>
 
-        <Sheet open={HelpIsOpen} onOpenChange={setHelpIsOpen} modal animation="medium" snapPoints={[88]}>
-          <Sheet.Overlay />
-          <Sheet.Frame bg={bgColor} borderRadius="$9">
-            <YStack bg={bgColor} px="$3" pb="$5" flex={1} >
-              <XStack ml="$2" mt="$3" gap="$2">
-                <H2 color={textColor1}>Help</H2>
-                <HelpCircle color={textColor1} mt="$1" alignSelf='center' size="$2" />
-                <XStack justifyContent="flex-end" f={1} mt="$2" mr="$2" gap="$5">
-                  <Pressable onPress={() => Linking.openURL('https://proofofpassport.com')}>
-                    <Image
-                      source={{ uri: Internet, width: 24, height: 24 }}
-                    />
-                  </Pressable>
-                  <Pressable onPress={() => Linking.openURL('https://t.me/proofofpassport')}>
-                    <Image
-                      source={{ uri: Telegram, width: 24, height: 24 }}
-                    />
-                  </Pressable>
-                  <Pressable onPress={() => Linking.openURL('https://x.com/proofofpassport')}>
-                    <Image
-                      source={{ uri: X, width: 24, height: 24 }}
-                    />
-                  </Pressable>
-                  <Pressable onPress={() => Linking.openURL('https://github.com/zk-passport/proof-of-passport')}>
-                    <Image
-                      tintColor={textColor1}
-                      source={{ uri: Github, width: 24, height: 24 }}
-                    />
-                  </Pressable>
+          <Sheet open={SettingsIsOpen} onOpenChange={setSettingsIsOpen} modal animation="medium" snapPoints={[88]}>
+            <Sheet.Overlay />
+            <Sheet.Frame bg={bgColor} borderRadius="$9">
+              <YStack p="$3" pb="$5" f={1} gap={height > 750 ? "$3" : "$1"} mb="$1.5">
+                <XStack gap="$2" ml="$2" >
+                  <H2 color={textColor1}>Settings</H2>
+                  <Cog color={textColor1} mt="$1" alignSelf='center' size="$2" />
                 </XStack>
-              </XStack>
-              <YStack flex={1} mt="$3" jc="space-evenly">
-                <YStack >
-                  <H3 color={textColor1}>Security and Privacy</H3>
-                  <Text color={textColor2} ml="$2" mt="$1">Proof of Passport uses zero-knowledge cryptography to allow you to prove facts about yourself like humanity, nationality or age without disclosing sensitive information.</Text>
-                </YStack>
-                <YStack >
-                  <H3 color={textColor1}>About ZK Proofs</H3>
-                  <Text color={textColor2} ml="$2" mt="$1">Zero-knowledge proofs rely on mathematical magic tricks to show the validity of some computation without revealing of all its inputs. In our case, the proof shows the passport has not been forged, but allows you to hide sensitive data.</Text>
-                </YStack>
-                <YStack gap="$1">
-                  <H3 color={textColor1}>FAQ</H3>
-                  <YStack ml="$1">
-                    <H4 color={textColor1}>Troubleshoot NFC scanning</H4>
-                    <Text color={textColor2} ml="$2" >Refer to <Text onPress={() => Linking.openURL('https://zk-passport.github.io/posts/how-to-scan-your-passport-using-nfc/')} color={blueColorLight} style={{ textDecorationLine: 'underline', fontStyle: 'italic' }}>this tutorial</Text> on how to scan your passport using NFC.</Text>
+                <Fieldset horizontal>
+                  <Label color={textColor1} width={225} justifyContent="flex-end" htmlFor="name" >
+                    Broken camera
+                  </Label>
+                  <Switch size="$3.5" checked={brokenCamera} onCheckedChange={setBrokenCamera}>
+                    <Switch.Thumb animation="bouncy" bc={bgColor} />
+                  </Switch>
+                </Fieldset>
+                {
+                  brokenCamera &&
+                  <YStack pl="$3" gap="$1">
+                    <Fieldset gap="$4" horizontal>
+                      <Label color={textColor1} width={160} justifyContent="flex-end" fontSize={13}>
+                        Passport Number
+                      </Label>
+                      <Input bg={componentBgColor} color={textColor1} h="$3.5" borderColor={passportNumber?.length === 9 ? "green" : "unset"} flex={1} id="passport_number" onChangeText={(text) => setPassportNumber(text.toUpperCase())} value={passportNumber} keyboardType="default" />
+                    </Fieldset>
+                    <Fieldset gap="$4" horizontal>
+                      <Label color={textColor1} width={160} justifyContent="flex-end" fontSize={13}>
+                        Date of birth (yymmdd)
+                      </Label>
+                      <Input bg={componentBgColor} color={textColor1} h="$3.5" borderColor={dateOfBirth?.length === 6 ? "green" : "unset"} flex={1} id="date_of_birth" onChangeText={setDateOfBirth} value={dateOfBirth} keyboardType={Platform.OS == "ios" ? "default" : "number-pad"} />
+                    </Fieldset>
+                    <Fieldset gap="$4" horizontal>
+                      <Label color={textColor1} width={160} justifyContent="flex-end" fontSize={13}>
+                        Date of expiry (yymmdd)
+                      </Label>
+                      <Input bg={componentBgColor} color={textColor1} h="$3.5" borderColor={dateOfExpiry?.length === 6 ? "green" : "unset"} flex={1} id="date_of_expiry" onChangeText={setDateOfExpiry} value={dateOfExpiry} keyboardType={Platform.OS == "ios" ? "default" : "number-pad"} />
+                    </Fieldset>
                   </YStack>
-                  <YStack ml="$1">
-                    <H4 color={textColor1}>My camera is down</H4>
-                    <Text color={textColor2} ml="$2">Go to settings and turn on the broken camera option.</Text>
-                  </YStack>
-                  <YStack ml="$1">
-                    <H4 color={textColor1}>My passport is not supported</H4>
-                    <Text color={textColor2} ml="$2">Please contact us on Telegram, or if you have programming skills, you can easily <Text onPress={() => Linking.openURL('https://t.me/proofofpassport')} color={blueColorLight} style={{ textDecorationLine: 'underline', fontStyle: 'italic' }}>contribute</Text> to the project by adding your signature algorithm.</Text>
-                  </YStack>
-                </YStack>
+                }
 
+                <Fieldset horizontal>
+                  <Label color={textColor1} width={225} justifyContent="flex-end" htmlFor="restart" >
+                    Private mode
+                  </Label>
+                  <Switch size="$3.5" checked={hideData} onCheckedChange={handleHideData}>
+                    <Switch.Thumb animation="bouncy" bc={bgColor} />
+                  </Switch>
+                </Fieldset>
+
+
+
+                <Fieldset gap="$4" mt="$1" horizontal>
+                  <Label color={textColor1} width={200} justifyContent="flex-end" htmlFor="restart">
+                    Restart to step 1
+                  </Label>
+                  <Button bg={componentBgColor} jc="center" borderColor={borderColor} borderWidth={1.2} size="$3.5" ml="$2" onPress={handleRestart}>
+                    <IterationCw color={textColor1} />
+                  </Button>
+                </Fieldset>
+
+                <Fieldset gap="$4" mt="$1" horizontal>
+                  <Label color={textColor1} width={200} justifyContent="flex-end" htmlFor="skip" >
+                    Use mock passport data
+                  </Label>
+                  <Button bg={componentBgColor} jc="center" borderColor={borderColor} borderWidth={1.2} size="$3.5" ml="$2" onPress={handleSkip}>
+                    <VenetianMask color={textColor1} />
+                  </Button>
+                </Fieldset>
+                <YStack flex={1} />
+
+                <YStack mb="$0">
+                  <Button p="$2.5" borderRadius="$3" bg={componentBgColor} jc="center" borderColor={borderColor} borderWidth={1.2} onPress={() => setSettingsIsOpen(false)} w="80%" alignSelf='center'>
+                    <Text color={textColor1} textAlign='center' fow="bold">Close</Text>
+                  </Button>
+                </YStack>
               </YStack>
-              {/* <YStack flex={1} jc="space-evenly">
+            </Sheet.Frame>
+          </Sheet>
 
-                <YStack >
-                  <H4 color={textColor1}>How do I scan my passport ?</H4>
-                  <Text color={textColor1}>1. Find the location of the NFC chip of your passport. Most of the time, it will be in the back cover. If you have an American passport, the front and back cover are NFC-protected, so you have to open your passport and scan the back cover from the inside.
-                    <Button pl="$1" unstyled h="$1" w="$3" jc="flex-end" onPress={() => Linking.openURL('https://zk-passport.github.io/posts/where-is-my-chip/')}>
-                      <ExternalLink color="#3185FC" size={12} />
-                    </Button>
-                  </Text>
-                  <Text color={textColor1} mt="$2">2. Find the location of the NFC reader of your phone. On an iPhone, it should be on the upper part of your phone. On Android phones, it should be in the center.
-                    <Button pl="$1" unstyled h="$1" w="$3" jc="flex-end" onPress={() => Linking.openURL('https://zk-passport.github.io/posts/locate-NFC-reader/')}>
-                      <ExternalLink color="#3185FC" size={12} />
-                    </Button>
-                  </Text>
-                  <Text color={textColor1} mt="$2">3. Keep your passport pressed against your phone when the NFC popup shows up and hold still.</Text>
-                </YStack>
-                <YStack gap="$1">
-                  <H4 color={textColor1} mt="$2">Security and Privacy</H4>
-                  <Text color={textColor1}>Proof of Passport uses zero-knowledge cryptography to allow you to prove facts about yourself like humanity, nationality or age without disclosing sensitive information. It works by generating a proof showing your passport data has been correctly signed by a government authority without revealing the signature.</Text>
-                </YStack>
-                <YStack gap="$2">
-                  <H4 color={textColor1} mt="$1">What are zero-knowledge proofs ?</H4>
-
-                  <Text color={textColor1}>Zero-knowledge proofs rely on mathematical magic tricks to show the correctness of some computation while hiding some inputs of its inputs. In our case, the proof shows the passport has not been forged, but allows you to hide sensitive data.</Text>
-                </YStack>
-
-                <YStack gap="$2">
-                  <H4 >Contacts</H4>
-                  <XStack mt="$2" ml="$3" gap="$5">
+          <Sheet open={HelpIsOpen} onOpenChange={setHelpIsOpen} modal animation="medium" snapPoints={[88]}>
+            <Sheet.Overlay />
+            <Sheet.Frame bg={bgColor} borderRadius="$9">
+              <YStack bg={bgColor} px="$3" pb="$5" flex={1} >
+                <XStack ml="$2" mt="$3" gap="$2">
+                  <H2 color={textColor1}>Help</H2>
+                  <HelpCircle color={textColor1} mt="$1" alignSelf='center' size="$2" />
+                  <XStack justifyContent="flex-end" f={1} mt="$2" mr="$2" gap="$5">
+                    <Pressable onPress={() => Linking.openURL('https://proofofpassport.com')}>
+                      <Image
+                        source={{ uri: Internet, width: 24, height: 24 }}
+                      />
+                    </Pressable>
                     <Pressable onPress={() => Linking.openURL('https://t.me/proofofpassport')}>
                       <Image
                         source={{ uri: Telegram, width: 24, height: 24 }}
@@ -382,71 +319,166 @@ const MainScreen: React.FC<MainScreenProps> = ({
                     </Pressable>
                     <Pressable onPress={() => Linking.openURL('https://github.com/zk-passport/proof-of-passport')}>
                       <Image
+                        tintColor={textColor1}
                         source={{ uri: Github, width: 24, height: 24 }}
                       />
                     </Pressable>
                   </XStack>
+                </XStack>
+                <YStack flex={1} mt="$3" jc="space-evenly">
+                  <YStack >
+                    <H3 color={textColor1}>Security and Privacy</H3>
+                    <Text color={textColor2} ml="$2" mt="$1">Proof of Passport uses zero-knowledge cryptography to allow you to prove facts about yourself like humanity, nationality or age without disclosing sensitive information.</Text>
+                  </YStack>
+                  <YStack >
+                    <H3 color={textColor1}>About ZK Proofs</H3>
+                    <Text color={textColor2} ml="$2" mt="$1">Zero-knowledge proofs rely on mathematical magic tricks to show the validity of some computation without revealing of all its inputs. In our case, the proof shows the passport has not been forged, but allows you to hide sensitive data.</Text>
+                  </YStack>
+                  <YStack gap="$1">
+                    <H3 color={textColor1}>FAQ</H3>
+                    <YStack ml="$1">
+                      <H4 color={textColor1}>Troubleshoot NFC scanning</H4>
+                      <Text color={textColor2} ml="$2" >Refer to <Text onPress={() => Linking.openURL('https://zk-passport.github.io/posts/how-to-scan-your-passport-using-nfc/')} color={blueColorLight} style={{ textDecorationLine: 'underline', fontStyle: 'italic' }}>this tutorial</Text> on how to scan your passport using NFC.</Text>
+                    </YStack>
+                    <YStack ml="$1">
+                      <H4 color={textColor1}>My camera is down</H4>
+                      <Text color={textColor2} ml="$2">Go to settings and turn on the broken camera option.</Text>
+                    </YStack>
+                    <YStack ml="$1">
+                      <H4 color={textColor1}>My passport is not supported</H4>
+                      <Text color={textColor2} ml="$2">Please contact us on Telegram, or if you have programming skills, you can easily <Text onPress={() => Linking.openURL('https://t.me/proofofpassport')} color={blueColorLight} style={{ textDecorationLine: 'underline', fontStyle: 'italic' }}>contribute</Text> to the project by adding your signature algorithm.</Text>
+                    </YStack>
+                  </YStack>
+
                 </YStack>
-              </YStack> */}
-              <Button mt="$3" bg={componentBgColor} jc="center" borderColor={borderColor} borderWidth={1.2} size="$3.5" ml="$2" alignSelf='center' w="80%" onPress={() => setHelpIsOpen(false)}>
-                <Text color={textColor1} w="80%" textAlign='center' fow="bold">Close</Text>
-              </Button>
+                {/* <YStack flex={1} jc="space-evenly">
 
-            </YStack>
-          </Sheet.Frame>
-        </Sheet>
-        <XStack bc="#343434" h={1.2} />
+                  <YStack >
+                    <H4 color={textColor1}>How do I scan my passport ?</H4>
+                    <Text color={textColor1}>1. Find the location of the NFC chip of your passport. Most of the time, it will be in the back cover. If you have an American passport, the front and back cover are NFC-protected, so you have to open your passport and scan the back cover from the inside.
+                      <Button pl="$1" unstyled h="$1" w="$3" jc="flex-end" onPress={() => Linking.openURL('https://zk-passport.github.io/posts/where-is-my-chip/')}>
+                        <ExternalLink color="#3185FC" size={12} />
+                      </Button>
+                    </Text>
+                    <Text color={textColor1} mt="$2">2. Find the location of the NFC reader of your phone. On an iPhone, it should be on the upper part of your phone. On Android phones, it should be in the center.
+                      <Button pl="$1" unstyled h="$1" w="$3" jc="flex-end" onPress={() => Linking.openURL('https://zk-passport.github.io/posts/locate-NFC-reader/')}>
+                        <ExternalLink color="#3185FC" size={12} />
+                      </Button>
+                    </Text>
+                    <Text color={textColor1} mt="$2">3. Keep your passport pressed against your phone when the NFC popup shows up and hold still.</Text>
+                  </YStack>
+                  <YStack gap="$1">
+                    <H4 color={textColor1} mt="$2">Security and Privacy</H4>
+                    <Text color={textColor1}>Proof of Passport uses zero-knowledge cryptography to allow you to prove facts about yourself like humanity, nationality or age without disclosing sensitive information. It works by generating a proof showing your passport data has been correctly signed by a government authority without revealing the signature.</Text>
+                  </YStack>
+                  <YStack gap="$2">
+                    <H4 color={textColor1} mt="$1">What are zero-knowledge proofs ?</H4>
+
+                    <Text color={textColor1}>Zero-knowledge proofs rely on mathematical magic tricks to show the correctness of some computation while hiding some inputs of its inputs. In our case, the proof shows the passport has not been forged, but allows you to hide sensitive data.</Text>
+                  </YStack>
+
+                  <YStack gap="$2">
+                    <H4 >Contacts</H4>
+                    <XStack mt="$2" ml="$3" gap="$5">
+                      <Pressable onPress={() => Linking.openURL('https://t.me/proofofpassport')}>
+                        <Image
+                          source={{ uri: Telegram, width: 24, height: 24 }}
+                        />
+                      </Pressable>
+                      <Pressable onPress={() => Linking.openURL('https://x.com/proofofpassport')}>
+                        <Image
+                          source={{ uri: X, width: 24, height: 24 }}
+                        />
+                      </Pressable>
+                      <Pressable onPress={() => Linking.openURL('https://github.com/zk-passport/proof-of-passport')}>
+                        <Image
+                          source={{ uri: Github, width: 24, height: 24 }}
+                        />
+                      </Pressable>
+                    </XStack>
+                  </YStack>
+                </YStack> */}
+                <Button mt="$3" bg={componentBgColor} jc="center" borderColor={borderColor} borderWidth={1.2} size="$3.5" ml="$2" alignSelf='center' w="80%" onPress={() => setHelpIsOpen(false)}>
+                  <Text color={textColor1} w="80%" textAlign='center' fow="bold">Close</Text>
+                </Button>
+
+              </YStack>
+            </Sheet.Frame>
+          </Sheet>
+          <XStack bc="#343434" h={1.2} />
+        </YStack>
+        <Tabs f={1} orientation="horizontal" flexDirection="column" defaultValue="scan" value={selectedTab} onValueChange={setSelectedTab}>
+          <ToastViewport flexDirection="column-reverse" top={15} right={0} left={0} />
+          <ToastMessage />
+          <Tabs.Content value="scan" f={1}>
+            <ScanScreen
+              onStartCameraScan={onStartCameraScan}
+              handleNFCScan={handleNFCScan}
+              step={step} />
+          </Tabs.Content>
+
+          <Tabs.Content value="app" f={1}>
+            <AppScreen
+              selectedApp={selectedApp}
+              setSelectedApp={setSelectedApp}
+              step={step}
+              setStep={setStep}
+              setSelectedTab={setSelectedTab}
+            />
+          </Tabs.Content>
+
+          <Tabs.Content value="prove" f={1}>
+            <ProveScreen
+              passportData={passportData}
+              disclosure={disclosure}
+              selectedApp={selectedApp}
+              handleDisclosureChange={handleDisclosureChange}
+              address={address}
+              setAddress={setAddress}
+              generatingProof={generatingProof}
+              handleProve={handleProve}
+              hideData={hideData}
+              ens={ens}
+              setEns={setEns}
+              majority={majority}
+              setMajority={setMajority}
+              zkeydownloadStatus={zkeydownloadStatus}
+            />
+          </Tabs.Content>
+          <Tabs.Content value="mint" f={1}>
+            <MintScreen
+              selectedApp={selectedApp}
+              step={step}
+              mintText={mintText}
+              proof={proof}
+              proofTime={proofTime}
+              handleMint={handleMint}
+            />
+          </Tabs.Content>
+        </Tabs>
       </YStack>
-      <Tabs f={1} orientation="horizontal" flexDirection="column" defaultValue="scan" value={selectedTab} onValueChange={setSelectedTab}>
-        <ToastViewport flexDirection="column-reverse" top={15} right={0} left={0} />
-        <ToastMessage />
-        <Tabs.Content value="scan" f={1}>
-          <ScanScreen
-            onStartCameraScan={onStartCameraScan}
-            handleNFCScan={handleNFCScan}
-            step={step} />
-        </Tabs.Content>
-
-        <Tabs.Content value="app" f={1}>
-          <AppScreen
-            selectedApp={selectedApp}
-            setSelectedApp={setSelectedApp}
-            step={step}
-            setStep={setStep}
-            setSelectedTab={setSelectedTab}
-          />
-        </Tabs.Content>
-
-        <Tabs.Content value="prove" f={1}>
-          <ProveScreen
-            passportData={passportData}
-            disclosure={disclosure}
-            selectedApp={selectedApp}
-            handleDisclosureChange={handleDisclosureChange}
-            address={address}
-            setAddress={setAddress}
-            generatingProof={generatingProof}
-            handleProve={handleProve}
-            hideData={hideData}
-            ens={ens}
-            setEns={setEns}
-            majority={majority}
-            setMajority={setMajority}
-            zkeydownloadStatus={zkeydownloadStatus}
-          />
-        </Tabs.Content>
-        <Tabs.Content value="mint" f={1}>
-          <MintScreen
-            selectedApp={selectedApp}
-            step={step}
-            mintText={mintText}
-            proof={proof}
-            proofTime={proofTime}
-            handleMint={handleMint}
-          />
-        </Tabs.Content>
-      </Tabs>
-    </YStack >
+      <Modal visible={showWarning} animationType="slide" transparent={true}>
+        <YStack bc="#161616" p={20} ai="center" jc="center" position="absolute" top={0} bottom={0} left={0} right={0}>
+          <YStack bc="#343434" p={20} borderRadius={10} ai="center" jc="center">
+            <Text fontWeight="bold" fontSize={18} color="#a0a0a0">👋 Hi</Text>
+            <Text mt={10} textAlign="center" color="#a0a0a0">
+              The app needs to download a large file (300MB). Please make sure you're connected to a Wi-Fi network before continuing.
+            </Text>
+            <XStack mt={20}>
+              <Button onPress={() => {
+                downloadZkey(
+                  setDownloadStatus,
+                  toast
+                );
+                setShowWarning(false)
+              }} bc="#4caf50" borderRadius={5} padding={10}>
+                <Text color="#ffffff">Continue</Text>
+              </Button>
+            </XStack>
+          </YStack>
+        </YStack>
+      </Modal>
+    </>
   );
 };
